[Fix] Update about me page
After considering the consequences of having disclosed dyslexia on there, I feel like it could affect my future employment opportunities. If somebody ever sees this commit (highly unlikely), please respect my decision to not include it there, and the fact that I kinda feel ashamed for it... I just don't want to be treated differentlly than everybody else. It was a dumb idea to include it there to begin with. I could rebase the git history and force push, but that doesn't feel right for some reason, idk. Let's just not talk about it, like ever, ok? Deal.